+++ frontend/src/main.ts
@@ -5,8 +5,11 @@ import { createApp } from 'vue'
 
 import App from './App.vue'
 import router from './router'
+import { setupErrorReporting } from './utils/error-reporting'
 
 const app = createApp(App)
+
+setupErrorReporting(app)
 
 app.use(createPinia())
 app.use(router)